fix: confidence data type

+++ src/oracle/interfaces/IOracleAdapter.sol
@@ -5,7 +5,7 @@ interface IOracleAdapter {
   function getLatestPrice(
     bytes32 _assetId,
     bool _isMax,
-    uint256 _confidenceThreshold
+    uint32 _confidenceThreshold
   ) external view returns (uint256, int32, uint256);
 
   function isSameAsset(bytes32 _assetId1, bytes32 _assetId2) external view returns (bool);
